+++ R/col_types.R
@@ -311,13 +311,13 @@ vroom_enquo <- function(x) {
   x
 }
 
-vroom_select <- function(x, col_select) {
+vroom_select <- function(x, col_select, id) {
   # reorder and rename columns
   if (inherits(col_select, "quosures") || !rlang::quo_is_null(col_select)) {
     if (inherits(col_select, "quosures")) {
-      vars <- tidyselect::vars_select(names(spec(x)$cols), !!!col_select)
+      vars <- tidyselect::vars_select(c(id, names(spec(x)$cols)), !!!col_select)
     } else {
-      vars <- tidyselect::vars_select(names(spec(x)$cols), !!col_select)
+      vars <- tidyselect::vars_select(c(id, names(spec(x)$cols)), !!col_select)
     }
     # This can't be just names(x) as we need to have skipped
     # names as well to pass to vars_select()
@@ -367,9 +367,9 @@ col_types_standardise <- function(col_types, col_names, col_select) {
 
   if (inherits(col_select, "quosures") || !rlang::quo_is_null(col_select)) {
     if (inherits(col_select, "quosures")) {
-      to_keep <- names(spec$cols) %in% tidyselect::vars_select(names(spec$cols), !!!col_select)
+      to_keep <- names(spec$cols) %in% tidyselect::vars_select(names(spec$cols), !!!col_select, .strict = FALSE)
     } else {
-      to_keep <- names(spec$cols) %in% tidyselect::vars_select(names(spec$cols), !!col_select)
+      to_keep <- names(spec$cols) %in% tidyselect::vars_select(names(spec$cols), !!col_select, .strict = FALSE)
     }
 
     spec$cols[!to_keep] <- rep(list(col_skip()), sum(!to_keep))